Refactor region from API bounding box into helper function
