+++ src/components/SharedComponents/Map/helpers/mapHelpers.ts
@@ -62,6 +62,32 @@ export function metersToLatitudeDelta( meters: number, latitude: number ): numbe
   return latitudeDelta;
 }
 
+interface TotalBounds {
+  nelat: number;
+  nelng: number;
+  swlat: number;
+  swlng: number;
+}
+export function getMapRegion( totalBounds: TotalBounds ) {
+  const {
+    nelat, nelng, swlat, swlng
+  } = totalBounds;
+  // Deltas shouldn't be negative
+  const latDelta = Math.abs( Number( nelat ) - Number( swlat ) );
+  const lngDelta = Math.abs( Number( nelng ) - Number( swlng ) );
+  const lat = nelat - ( latDelta / 2 );
+  const lng = nelng - ( lngDelta / 2 );
+
+  return {
+    latitude: lat,
+    longitude: lng,
+    // Pad the detlas so the user sees the full range, make sure we don't
+    // specify impossible deltas like 190 degrees of latitude
+    latitudeDelta: Math.min( latDelta + 5, 175 ),
+    longitudeDelta: Math.min( lngDelta + 5, 355 )
+  };
+}
+
 interface Params {
   [key: string]: unknown;
 }
